Move host fs impl into virtual fs crate
