+++ lib/virtual-fs/Cargo.toml
@@ -7,7 +7,8 @@ license = "MIT"
 edition = "2018"
 
 [dependencies]
+libc = { version = "^0.2", default-features = false }
 thiserror = "1"
 tracing = { version = "0.1" }
 typetag = "0.1"
-#serde = { version = "1.0", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"] }
