+++ lib/wasi/src/state/builder.rs
@@ -5,7 +5,7 @@ use crate::syscalls::types::{__WASI_STDERR_FILENO, __WASI_STDIN_FILENO, __WASI_S
 use crate::WasiEnv;
 use std::path::{Path, PathBuf};
 use thiserror::Error;
-use wasmer_virtual_fs::{FsError, WasiFile};
+use wasmer_virtual_fs::{FsError, VirtualFile};
 
 /// Creates an empty [`WasiStateBuilder`].
 ///
@@ -40,9 +40,9 @@ pub struct WasiStateBuilder {
     preopens: Vec<PreopenedDir>,
     #[allow(clippy::type_complexity)]
     setup_fs_fn: Option<Box<dyn Fn(&mut WasiFs) -> Result<(), String> + Send>>,
-    stdout_override: Option<Box<dyn WasiFile>>,
-    stderr_override: Option<Box<dyn WasiFile>>,
-    stdin_override: Option<Box<dyn WasiFile>>,
+    stdout_override: Option<Box<dyn VirtualFile>>,
+    stderr_override: Option<Box<dyn VirtualFile>>,
+    stdin_override: Option<Box<dyn VirtualFile>>,
 }
 
 impl std::fmt::Debug for WasiStateBuilder {
@@ -261,7 +261,7 @@ impl WasiStateBuilder {
 
     /// Overwrite the default WASI `stdout`, if you want to hold on to the
     /// original `stdout` use [`WasiFs::swap_file`] after building.
-    pub fn stdout(&mut self, new_file: Box<dyn WasiFile>) -> &mut Self {
+    pub fn stdout(&mut self, new_file: Box<dyn VirtualFile>) -> &mut Self {
         self.stdout_override = Some(new_file);
 
         self
@@ -269,7 +269,7 @@ impl WasiStateBuilder {
 
     /// Overwrite the default WASI `stderr`, if you want to hold on to the
     /// original `stderr` use [`WasiFs::swap_file`] after building.
-    pub fn stderr(&mut self, new_file: Box<dyn WasiFile>) -> &mut Self {
+    pub fn stderr(&mut self, new_file: Box<dyn VirtualFile>) -> &mut Self {
         self.stderr_override = Some(new_file);
 
         self
@@ -277,7 +277,7 @@ impl WasiStateBuilder {
 
     /// Overwrite the default WASI `stdin`, if you want to hold on to the
     /// original `stdin` use [`WasiFs::swap_file`] after building.
-    pub fn stdin(&mut self, new_file: Box<dyn WasiFile>) -> &mut Self {
+    pub fn stdin(&mut self, new_file: Box<dyn VirtualFile>) -> &mut Self {
         self.stdin_override = Some(new_file);
 
         self
@@ -398,6 +398,7 @@ impl WasiStateBuilder {
                     env
                 })
                 .collect(),
+            fs_backing: Box::new(wasmer_virtual_fs::host_fs::HostFileSystem),
         })
     }
