+++ lib/wasi/src/state/types.rs
@@ -1,17 +1,14 @@
 /// types for use in the WASI filesystem
 use crate::syscalls::types::*;
-use serde::{de, Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 #[cfg(unix)]
 use std::convert::TryInto;
 use std::{
     collections::VecDeque,
-    fs,
     io::{self, Read, Seek, Write},
-    path::PathBuf,
-    time::SystemTime,
 };
-use tracing::debug;
-use wasmer_virtual_fs::{FsError, WasiFile};
+pub use wasmer_virtual_fs::host_fs::{HostFile, Stderr, Stdin, Stdout};
+use wasmer_virtual_fs::{FsError, VirtualFile};
 
 pub fn fs_error_from_wasi_err(err: __wasi_errno_t) -> FsError {
     match err {
@@ -182,7 +179,7 @@ impl PollEventBuilder {
 
 #[cfg(unix)]
 pub(crate) fn poll(
-    selfs: &[&dyn WasiFile],
+    selfs: &[&dyn VirtualFile],
     events: &[PollEventSet],
     seen_events: &mut [PollEventSet],
 ) -> Result<u32, FsError> {
@@ -215,553 +212,16 @@ pub(crate) fn poll(
 
 #[cfg(not(unix))]
 pub(crate) fn poll(
-    _selfs: &[&dyn WasiFile],
+    _selfs: &[&dyn VirtualFile],
     _events: &[PollEventSet],
     _seen_events: &mut [PollEventSet],
 ) -> Result<(), FsError> {
-    unimplemented!("HostFile::poll in WasiFile is not implemented for non-Unix-like targets yet");
+    unimplemented!(
+        "HostFile::poll in VirtualFile is not implemented for non-Unix-like targets yet"
+    );
 }
 
 pub trait WasiPath {}
-
-/// A thin wrapper around `std::fs::File`
-#[derive(Debug, Serialize)]
-pub struct HostFile {
-    #[serde(skip_serializing)]
-    pub inner: fs::File,
-    pub host_path: PathBuf,
-    flags: u16,
-}
-
-impl<'de> Deserialize<'de> for HostFile {
-    fn deserialize<D>(deserializer: D) -> Result<HostFile, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(field_identifier, rename_all = "snake_case")]
-        enum Field {
-            HostPath,
-            Flags,
-        }
-
-        struct HostFileVisitor;
-
-        impl<'de> de::Visitor<'de> for HostFileVisitor {
-            type Value = HostFile;
-
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("struct HostFile")
-            }
-
-            fn visit_seq<V>(self, mut seq: V) -> Result<Self::Value, V::Error>
-            where
-                V: de::SeqAccess<'de>,
-            {
-                let host_path = seq
-                    .next_element()?
-                    .ok_or_else(|| de::Error::invalid_length(0, &self))?;
-                let flags = seq
-                    .next_element()?
-                    .ok_or_else(|| de::Error::invalid_length(1, &self))?;
-                let inner = std::fs::OpenOptions::new()
-                    .read(flags & HostFile::READ != 0)
-                    .write(flags & HostFile::WRITE != 0)
-                    .append(flags & HostFile::APPEND != 0)
-                    .open(&host_path)
-                    .map_err(|_| de::Error::custom("Could not open file on this system"))?;
-                Ok(HostFile {
-                    inner,
-                    host_path,
-                    flags,
-                })
-            }
-
-            fn visit_map<V>(self, mut map: V) -> Result<Self::Value, V::Error>
-            where
-                V: de::MapAccess<'de>,
-            {
-                let mut host_path = None;
-                let mut flags = None;
-                while let Some(key) = map.next_key()? {
-                    match key {
-                        Field::HostPath => {
-                            if host_path.is_some() {
-                                return Err(de::Error::duplicate_field("host_path"));
-                            }
-                            host_path = Some(map.next_value()?);
-                        }
-                        Field::Flags => {
-                            if flags.is_some() {
-                                return Err(de::Error::duplicate_field("flags"));
-                            }
-                            flags = Some(map.next_value()?);
-                        }
-                    }
-                }
-                let host_path = host_path.ok_or_else(|| de::Error::missing_field("host_path"))?;
-                let flags = flags.ok_or_else(|| de::Error::missing_field("flags"))?;
-                let inner = std::fs::OpenOptions::new()
-                    .read(flags & HostFile::READ != 0)
-                    .write(flags & HostFile::WRITE != 0)
-                    .append(flags & HostFile::APPEND != 0)
-                    .open(&host_path)
-                    .map_err(|_| de::Error::custom("Could not open file on this system"))?;
-                Ok(HostFile {
-                    inner,
-                    host_path,
-                    flags,
-                })
-            }
-        }
-
-        const FIELDS: &[&str] = &["host_path", "flags"];
-        deserializer.deserialize_struct("HostFile", FIELDS, HostFileVisitor)
-    }
-}
-
-impl HostFile {
-    const READ: u16 = 1;
-    const WRITE: u16 = 2;
-    const APPEND: u16 = 4;
-
-    /// creates a new host file from a `std::fs::File` and a path
-    pub fn new(file: fs::File, host_path: PathBuf, read: bool, write: bool, append: bool) -> Self {
-        let mut flags = 0;
-        if read {
-            flags |= Self::READ;
-        }
-        if write {
-            flags |= Self::WRITE;
-        }
-        if append {
-            flags |= Self::APPEND;
-        }
-        Self {
-            inner: file,
-            host_path,
-            flags,
-        }
-    }
-
-    pub fn metadata(&self) -> fs::Metadata {
-        self.inner.metadata().unwrap()
-    }
-}
-
-impl Read for HostFile {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.inner.read(buf)
-    }
-    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> io::Result<usize> {
-        self.inner.read_to_end(buf)
-    }
-    fn read_to_string(&mut self, buf: &mut String) -> io::Result<usize> {
-        self.inner.read_to_string(buf)
-    }
-    fn read_exact(&mut self, buf: &mut [u8]) -> io::Result<()> {
-        self.inner.read_exact(buf)
-    }
-}
-impl Seek for HostFile {
-    fn seek(&mut self, pos: io::SeekFrom) -> io::Result<u64> {
-        self.inner.seek(pos)
-    }
-}
-impl Write for HostFile {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.inner.write(buf)
-    }
-    fn flush(&mut self) -> io::Result<()> {
-        self.inner.flush()
-    }
-    fn write_all(&mut self, buf: &[u8]) -> io::Result<()> {
-        self.inner.write_all(buf)
-    }
-    fn write_fmt(&mut self, fmt: ::std::fmt::Arguments) -> io::Result<()> {
-        self.inner.write_fmt(fmt)
-    }
-}
-
-#[typetag::serde]
-impl WasiFile for HostFile {
-    fn last_accessed(&self) -> u64 {
-        self.metadata()
-            .accessed()
-            .ok()
-            .and_then(|ct| ct.duration_since(SystemTime::UNIX_EPOCH).ok())
-            .map(|ct| ct.as_nanos() as u64)
-            .unwrap_or(0)
-    }
-
-    fn set_last_accessed(&self, _last_accessed: __wasi_timestamp_t) {
-        // TODO: figure out how to do this
-    }
-
-    fn last_modified(&self) -> u64 {
-        self.metadata()
-            .modified()
-            .ok()
-            .and_then(|ct| ct.duration_since(SystemTime::UNIX_EPOCH).ok())
-            .map(|ct| ct.as_nanos() as u64)
-            .unwrap_or(0)
-    }
-
-    fn set_last_modified(&self, _last_modified: __wasi_timestamp_t) {
-        // TODO: figure out how to do this
-    }
-
-    fn created_time(&self) -> u64 {
-        self.metadata()
-            .created()
-            .ok()
-            .and_then(|ct| ct.duration_since(SystemTime::UNIX_EPOCH).ok())
-            .map(|ct| ct.as_nanos() as u64)
-            .unwrap_or(0)
-    }
-
-    fn set_created_time(&self, _created_time: __wasi_timestamp_t) {
-        // TODO: figure out how to do this
-    }
-
-    fn size(&self) -> u64 {
-        self.metadata().len()
-    }
-
-    fn set_len(&mut self, new_size: __wasi_filesize_t) -> Result<(), FsError> {
-        fs::File::set_len(&self.inner, new_size).map_err(Into::into)
-    }
-
-    fn unlink(&mut self) -> Result<(), FsError> {
-        std::fs::remove_file(&self.host_path).map_err(Into::into)
-    }
-    fn sync_to_disk(&self) -> Result<(), FsError> {
-        self.inner.sync_all().map_err(Into::into)
-    }
-
-    fn rename_file(&self, new_name: &std::path::Path) -> Result<(), FsError> {
-        std::fs::rename(&self.host_path, new_name).map_err(Into::into)
-    }
-
-    fn bytes_available(&self) -> Result<usize, FsError> {
-        // unwrap is safe because of get_raw_fd implementation
-        let host_fd = self.get_raw_fd().unwrap();
-
-        host_file_bytes_available(host_fd)
-    }
-
-    #[cfg(unix)]
-    fn get_raw_fd(&self) -> Option<i32> {
-        use std::os::unix::io::AsRawFd;
-        Some(self.inner.as_raw_fd())
-    }
-    #[cfg(not(unix))]
-    fn get_raw_fd(&self) -> Option<i32> {
-        unimplemented!(
-            "HostFile::get_raw_fd in WasiFile is not implemented for non-Unix-like targets yet"
-        );
-    }
-}
-
-#[cfg(unix)]
-fn host_file_bytes_available(host_fd: i32) -> Result<usize, FsError> {
-    let mut bytes_found = 0 as libc::c_int;
-    let result = unsafe { libc::ioctl(host_fd, libc::FIONREAD, &mut bytes_found) };
-
-    match result {
-        // success
-        0 => Ok(bytes_found.try_into().unwrap_or(0)),
-        libc::EBADF => Err(FsError::InvalidFd),
-        libc::EFAULT => Err(FsError::InvalidData),
-        libc::EINVAL => Err(FsError::InvalidInput),
-        _ => Err(FsError::IOError),
-    }
-}
-
-#[cfg(not(unix))]
-fn host_file_bytes_available(_raw_fd: i32) -> Result<usize, FsError> {
-    unimplemented!("host_file_bytes_available not yet implemented for non-Unix-like targets.  This probably means the program tried to use wasi::poll_oneoff")
-}
-
-/// A wrapper type around Stdout that implements `WasiFile` and
-/// `Serialize` + `Deserialize`.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Stdout;
-impl Read for Stdout {
-    fn read(&mut self, _buf: &mut [u8]) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stdout",
-        ))
-    }
-    fn read_to_end(&mut self, _buf: &mut Vec<u8>) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stdout",
-        ))
-    }
-    fn read_to_string(&mut self, _buf: &mut String) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stdout",
-        ))
-    }
-    fn read_exact(&mut self, _buf: &mut [u8]) -> io::Result<()> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stdout",
-        ))
-    }
-}
-impl Seek for Stdout {
-    fn seek(&mut self, _pos: io::SeekFrom) -> io::Result<u64> {
-        Err(io::Error::new(io::ErrorKind::Other, "can not seek stdout"))
-    }
-}
-impl Write for Stdout {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        io::stdout().write(buf)
-    }
-    fn flush(&mut self) -> io::Result<()> {
-        io::stdout().flush()
-    }
-    fn write_all(&mut self, buf: &[u8]) -> io::Result<()> {
-        io::stdout().write_all(buf)
-    }
-    fn write_fmt(&mut self, fmt: ::std::fmt::Arguments) -> io::Result<()> {
-        io::stdout().write_fmt(fmt)
-    }
-}
-
-#[typetag::serde]
-impl WasiFile for Stdout {
-    fn last_accessed(&self) -> u64 {
-        0
-    }
-    fn last_modified(&self) -> u64 {
-        0
-    }
-    fn created_time(&self) -> u64 {
-        0
-    }
-    fn size(&self) -> u64 {
-        0
-    }
-    fn set_len(&mut self, _new_size: __wasi_filesize_t) -> Result<(), FsError> {
-        debug!("Calling WasiFile::set_len on stdout; this is probably a bug");
-        Err(FsError::PermissionDenied)
-    }
-    fn unlink(&mut self) -> Result<(), FsError> {
-        Ok(())
-    }
-
-    fn bytes_available(&self) -> Result<usize, FsError> {
-        // unwrap is safe because of get_raw_fd implementation
-        let host_fd = self.get_raw_fd().unwrap();
-
-        host_file_bytes_available(host_fd)
-    }
-
-    #[cfg(unix)]
-    fn get_raw_fd(&self) -> Option<i32> {
-        use std::os::unix::io::AsRawFd;
-        Some(io::stdout().as_raw_fd())
-    }
-
-    #[cfg(not(unix))]
-    fn get_raw_fd(&self) -> Option<i32> {
-        unimplemented!(
-            "Stdout::get_raw_fd in WasiFile is not implemented for non-Unix-like targets yet"
-        );
-    }
-}
-
-/// A wrapper type around Stderr that implements `WasiFile` and
-/// `Serialize` + `Deserialize`.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Stderr;
-impl Read for Stderr {
-    fn read(&mut self, _buf: &mut [u8]) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stderr",
-        ))
-    }
-    fn read_to_end(&mut self, _buf: &mut Vec<u8>) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stderr",
-        ))
-    }
-    fn read_to_string(&mut self, _buf: &mut String) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stderr",
-        ))
-    }
-    fn read_exact(&mut self, _buf: &mut [u8]) -> io::Result<()> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not read from stderr",
-        ))
-    }
-}
-impl Seek for Stderr {
-    fn seek(&mut self, _pos: io::SeekFrom) -> io::Result<u64> {
-        Err(io::Error::new(io::ErrorKind::Other, "can not seek stderr"))
-    }
-}
-impl Write for Stderr {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        io::stderr().write(buf)
-    }
-    fn flush(&mut self) -> io::Result<()> {
-        io::stderr().flush()
-    }
-    fn write_all(&mut self, buf: &[u8]) -> io::Result<()> {
-        io::stderr().write_all(buf)
-    }
-    fn write_fmt(&mut self, fmt: ::std::fmt::Arguments) -> io::Result<()> {
-        io::stderr().write_fmt(fmt)
-    }
-}
-
-#[typetag::serde]
-impl WasiFile for Stderr {
-    fn last_accessed(&self) -> u64 {
-        0
-    }
-    fn last_modified(&self) -> u64 {
-        0
-    }
-    fn created_time(&self) -> u64 {
-        0
-    }
-    fn size(&self) -> u64 {
-        0
-    }
-    fn set_len(&mut self, _new_size: __wasi_filesize_t) -> Result<(), FsError> {
-        debug!("Calling WasiFile::set_len on stderr; this is probably a bug");
-        Err(FsError::PermissionDenied)
-    }
-    fn unlink(&mut self) -> Result<(), FsError> {
-        Ok(())
-    }
-
-    fn bytes_available(&self) -> Result<usize, FsError> {
-        // unwrap is safe because of get_raw_fd implementation
-        let host_fd = self.get_raw_fd().unwrap();
-
-        host_file_bytes_available(host_fd)
-    }
-
-    #[cfg(unix)]
-    fn get_raw_fd(&self) -> Option<i32> {
-        use std::os::unix::io::AsRawFd;
-        Some(io::stderr().as_raw_fd())
-    }
-
-    #[cfg(not(unix))]
-    fn get_raw_fd(&self) -> Option<i32> {
-        unimplemented!(
-            "Stderr::get_raw_fd in WasiFile is not implemented for non-Unix-like targets yet"
-        );
-    }
-}
-
-/// A wrapper type around Stdin that implements `WasiFile` and
-/// `Serialize` + `Deserialize`.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Stdin;
-impl Read for Stdin {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        io::stdin().read(buf)
-    }
-    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> io::Result<usize> {
-        io::stdin().read_to_end(buf)
-    }
-    fn read_to_string(&mut self, buf: &mut String) -> io::Result<usize> {
-        io::stdin().read_to_string(buf)
-    }
-    fn read_exact(&mut self, buf: &mut [u8]) -> io::Result<()> {
-        io::stdin().read_exact(buf)
-    }
-}
-impl Seek for Stdin {
-    fn seek(&mut self, _pos: io::SeekFrom) -> io::Result<u64> {
-        Err(io::Error::new(io::ErrorKind::Other, "can not seek stdin"))
-    }
-}
-impl Write for Stdin {
-    fn write(&mut self, _buf: &[u8]) -> io::Result<usize> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not write to stdin",
-        ))
-    }
-    fn flush(&mut self) -> io::Result<()> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not write to stdin",
-        ))
-    }
-    fn write_all(&mut self, _buf: &[u8]) -> io::Result<()> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not write to stdin",
-        ))
-    }
-    fn write_fmt(&mut self, _fmt: ::std::fmt::Arguments) -> io::Result<()> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "can not write to stdin",
-        ))
-    }
-}
-
-#[typetag::serde]
-impl WasiFile for Stdin {
-    fn last_accessed(&self) -> u64 {
-        0
-    }
-    fn last_modified(&self) -> u64 {
-        0
-    }
-    fn created_time(&self) -> u64 {
-        0
-    }
-    fn size(&self) -> u64 {
-        0
-    }
-    fn set_len(&mut self, _new_size: __wasi_filesize_t) -> Result<(), FsError> {
-        debug!("Calling WasiFile::set_len on stdin; this is probably a bug");
-        Err(FsError::PermissionDenied)
-    }
-
-    fn unlink(&mut self) -> Result<(), FsError> {
-        Ok(())
-    }
-
-    fn bytes_available(&self) -> Result<usize, FsError> {
-        // unwrap is safe because of get_raw_fd implementation
-        let host_fd = self.get_raw_fd().unwrap();
-
-        host_file_bytes_available(host_fd)
-    }
-
-    #[cfg(unix)]
-    fn get_raw_fd(&self) -> Option<i32> {
-        use std::os::unix::io::AsRawFd;
-        Some(io::stdin().as_raw_fd())
-    }
-
-    #[cfg(not(unix))]
-    fn get_raw_fd(&self) -> Option<i32> {
-        unimplemented!(
-            "Stdin::get_raw_fd in WasiFile is not implemented for non-Unix-like targets yet"
-        );
-    }
-}
 
 /// For piping stdio. Stores all output / input in a byte-vector.
 #[derive(Debug, Default, Serialize, Deserialize)]
@@ -805,7 +265,7 @@ impl Seek for Pipe {
 }
 
 #[typetag::serde]
-impl WasiFile for Pipe {
+impl VirtualFile for Pipe {
     fn last_accessed(&self) -> u64 {
         0
     }
